requirements.txt
openai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pillow>=10.0.0
flask>=3.0.0
werkzeug>=3.0.0
pymupdf>=1.23.0
numpy>=1.24.0
flask-login>=0.6.3
flask-sqlalchemy>=3.0.0
flask-wtf>=1.1.0
bcrypt>=4.0.0
stripe>=7.0.0
flask-limiter>=3.5.0
email-validator>=2.0.0